abstraction-level-generator/backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
aiohttp==3.7.4
pydantic==1.8.2
python-dotenv==0.19.0
